+++ wazo_calld/plugins/calls/schemas.py
@@ -69,7 +69,6 @@ class CallSchema(CallBaseSchema):
     is_caller = fields.Boolean()
     dialed_extension = fields.String()
     sip_call_id = fields.String()
-    channel_name = fields.String()
 
     @post_dump()
     def default_peer_caller_id_number(self, call):